+++ node-matchers/src/main/java/com/mattunderscore/trees/matchers/IdentityMatcher.java
@@ -25,17 +25,18 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 
 package com.mattunderscore.trees.matchers;
 
-import com.mattunderscore.trees.tree.OpenNode;
-import com.mattunderscore.trees.selection.NodeMatcher;
+import java.util.function.Predicate;
 
 import net.jcip.annotations.Immutable;
+
+import com.mattunderscore.trees.tree.OpenNode;
 
 /**
  * Matches nodes to the same element passed to it.
  * @author Matt Champion on 16/08/14.
  */
 @Immutable
-public final class IdentityMatcher<E> implements NodeMatcher<E> {
+public final class IdentityMatcher<E> implements Predicate<OpenNode<? extends E, ?>> {
     private final E value;
 
     public IdentityMatcher(E value) {
@@ -43,7 +44,7 @@ public final class IdentityMatcher<E> implements NodeMatcher<E> {
     }
 
     @Override
-    public boolean matches(OpenNode<? extends E, ?> node) {
+    public boolean test(OpenNode<? extends E, ?> node) {
         return node.getElement() == value;
     }