+++ node-matchers/src/main/java/com/mattunderscore/trees/matchers/NeverMatcher.java
@@ -25,21 +25,22 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 
 package com.mattunderscore.trees.matchers;
 
-import com.mattunderscore.trees.tree.OpenNode;
-import com.mattunderscore.trees.selection.NodeMatcher;
+import java.util.function.Predicate;
 
 import net.jcip.annotations.Immutable;
+
+import com.mattunderscore.trees.tree.OpenNode;
 
 /**
  * Matches no nodes.
  * @author Matt Champion on 16/08/14.
  */
 @Immutable
-public final class NeverMatcher<E> implements NodeMatcher<E> {
-    private static final NodeMatcher<?> INSTANCE = new NeverMatcher<>();
+public final class NeverMatcher<E> implements Predicate<OpenNode<? extends E, ?>> {
+    private static final Predicate INSTANCE = new NeverMatcher<>();
 
     @Override
-    public boolean matches(OpenNode<? extends E, ?> node) {
+    public boolean test(OpenNode<? extends E, ?> node) {
         return false;
     }
 
@@ -54,7 +55,7 @@ public final class NeverMatcher<E> implements NodeMatcher<E> {
     }
 
     @SuppressWarnings("unchecked")
-    public static <E> NodeMatcher<E> create() {
-        return (NodeMatcher<E>)INSTANCE;
+    public static <E> Predicate<OpenNode<? extends E, ?>> create() {
+        return (Predicate<OpenNode<? extends E, ?>>)INSTANCE;
     }
 }