+++ node-matchers/src/test/java/com/mattunderscore/trees/matchers/EqualityMatcherTest.java
@@ -29,10 +29,11 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
+import java.util.function.Predicate;
+
 import org.junit.Test;
 
 import com.mattunderscore.trees.base.ImmutableNode;
-import com.mattunderscore.trees.selection.NodeMatcher;
 import com.mattunderscore.trees.tree.Node;
 import com.mattunderscore.trees.tree.OpenNode;
 
@@ -45,21 +46,21 @@ public final class EqualityMatcherTest {
     @Test
     public void matches() {
         final OpenNode<String, Node<String>> node = new ImmutableNode<String, Node<String>>("a", new Object[0]) {};
-        final NodeMatcher<String> matcher = new EqualityMatcher<>("a");
-        assertTrue(matcher.matches(node));
+        final Predicate<OpenNode<? extends String, ?>> matcher = new EqualityMatcher<>("a");
+        assertTrue(matcher.test(node));
     }
 
     @Test
     public void notMatches() {
         final OpenNode<String, Node<String>> node = new ImmutableNode<String, Node<String>>("a", new Object[0]) {};
-        final NodeMatcher<String> matcher = new EqualityMatcher<>("b");
-        assertFalse(matcher.matches(node));
+        final Predicate<OpenNode<? extends String, ?>> matcher = new EqualityMatcher<>("b");
+        assertFalse(matcher.test(node));
     }
 
     @Test
     public void testEquals() {
-        final NodeMatcher<String> matcher0 = new EqualityMatcher<>("a");
-        final NodeMatcher<String> matcher1 = new EqualityMatcher<>("a");
+        final Predicate<OpenNode<? extends String, ?>> matcher0 = new EqualityMatcher<>("a");
+        final Predicate<OpenNode<? extends String, ?>> matcher1 = new EqualityMatcher<>("a");
 
         assertTrue(matcher0.equals(matcher1));
         assertTrue(matcher1.equals(matcher0));
@@ -68,22 +69,22 @@ public final class EqualityMatcherTest {
 
     @Test
     public void testNotEquals0() {
-        final NodeMatcher<String> matcher0 = new EqualityMatcher<>("a");
-        final NodeMatcher<String> matcher1 = new EqualityMatcher<>("b");
+        final Predicate<OpenNode<? extends String, ?>> matcher0 = new EqualityMatcher<>("a");
+        final Predicate<OpenNode<? extends String, ?>> matcher1 = new EqualityMatcher<>("b");
 
         assertFalse(matcher0.equals(matcher1));
     }
 
     @Test
     public void testNotEquals1() {
-        final NodeMatcher<String> matcher0 = new EqualityMatcher<>("a");
+        final Predicate<OpenNode<? extends String, ?>> matcher0 = new EqualityMatcher<>("a");
 
         assertFalse(matcher0.equals(null));
     }
 
     @Test
     public void testNotEquals2() {
-        final NodeMatcher<String> matcher0 = new EqualityMatcher<>("a");
+        final Predicate<OpenNode<? extends String, ?>> matcher0 = new EqualityMatcher<>("a");
 
         assertFalse(matcher0.equals(new Object()));
     }