+++ node-matchers/src/test/java/com/mattunderscore/trees/matchers/IdentityMatcherTest.java
@@ -29,10 +29,11 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
+import java.util.function.Predicate;
+
 import org.junit.Test;
 
 import com.mattunderscore.trees.base.ImmutableNode;
-import com.mattunderscore.trees.selection.NodeMatcher;
 import com.mattunderscore.trees.tree.Node;
 import com.mattunderscore.trees.tree.OpenNode;
 
@@ -45,22 +46,22 @@ public final class IdentityMatcherTest {
     public void matches() {
         final String element = "a";
         final OpenNode<String, Node<String>> node = new ImmutableNode<String, Node<String>>(element, new Object[0]) {};
-        final NodeMatcher<String> matcher = new IdentityMatcher<>(element);
-        assertTrue(matcher.matches(node));
+        final Predicate<OpenNode<? extends String, ?>> matcher = new IdentityMatcher<>(element);
+        assertTrue(matcher.test(node));
     }
 
     @Test
     public void notMatches() {
         final OpenNode<String, Node<String>> node = new ImmutableNode<String, Node<String>>(new String("a"), new Object[0]) {};
-        final NodeMatcher<String> matcher = new IdentityMatcher<>(new String("a"));
-        assertFalse(matcher.matches(node));
+        final Predicate<OpenNode<? extends String, ?>> matcher = new IdentityMatcher<>(new String("a"));
+        assertFalse(matcher.test(node));
     }
 
     @Test
     public void testEquals() {
         final String element = "a";
-        final NodeMatcher<String> matcher0 = new IdentityMatcher<>(element);
-        final NodeMatcher<String> matcher1 = new IdentityMatcher<>(element);
+        final Predicate<OpenNode<? extends String, ?>> matcher0 = new IdentityMatcher<>(element);
+        final Predicate<OpenNode<? extends String, ?>> matcher1 = new IdentityMatcher<>(element);
 
         assertTrue(matcher0.equals(matcher1));
         assertTrue(matcher1.equals(matcher0));
@@ -69,22 +70,22 @@ public final class IdentityMatcherTest {
 
     @Test
     public void testNotEquals0() {
-        final NodeMatcher<String> matcher0 = new IdentityMatcher<>(new String("a"));
-        final NodeMatcher<String> matcher1 = new IdentityMatcher<>(new String("a"));
+        final Predicate<OpenNode<? extends String, ?>> matcher0 = new IdentityMatcher<>(new String("a"));
+        final Predicate<OpenNode<? extends String, ?>> matcher1 = new IdentityMatcher<>(new String("a"));
 
         assertFalse(matcher0.equals(matcher1));
     }
 
     @Test
     public void testNotEquals1() {
-        final NodeMatcher matcher0 = new IdentityMatcher<>(new String("a"));
+        final Predicate<OpenNode<? extends String, ?>> matcher0 = new IdentityMatcher<>(new String("a"));
 
         assertFalse(matcher0.equals(null));
     }
 
     @Test
     public void testNotEquals2() {
-        final NodeMatcher matcher0 = new IdentityMatcher<>(new String("a"));
+        final Predicate<OpenNode<? extends String, ?>> matcher0 = new IdentityMatcher<>(new String("a"));
 
         assertFalse(matcher0.equals(new Object()));
     }